Refactoring in preparation for a change....
When the parser reaches an "Or", if one of the choices is to repeat something, that choice has priority over the other branches.  I already do this using an Expression called "Fallback".

Going forward I will need to add another distinction- If part of a sequence is optional (ie- it ends with a "?" in the grammar), this will take priority over anything else.

I have removed the expression "Fallback", and replaced it with a "Priority" construction, which can be either Low, Medium, or High.

No changes to the logic have been made yet, so options do not yet have priority over anything else.  This will be done in a future commit.